remove benchmarks
